+++ crypto/src/main/rust/latticegadget.rs
@@ -43,10 +43,10 @@ fn decompose_slice<Z: IntegerModRing, R: PolynomialRing<Z> + Clone>(
     slice: &[R],
     radix_mask: <Z::Int as Integer>::Limb,
     radix_shift: <Z::Int as Integer>::Limb,
-    digits: usize,
+    digits: u32,
 ) -> Vec<R> {
-    let mut pieces = vec![R::ZERO; slice.len() * digits];
-    for (polynomial, pieces) in zip(slice, pieces.chunks_exact_mut(digits)) {
+    let mut pieces = vec![R::ZERO; slice.len() * digits as usize];
+    for (polynomial, pieces) in zip(slice, pieces.chunks_exact_mut(digits as usize)) {
         decompose_impl(polynomial, radix_mask, radix_shift, pieces)
     }
     pieces
@@ -56,10 +56,10 @@ pub fn decompose_integer<Z: IntegerModRing>(
     integer: &Z,
     radix_mask: <Z::Int as Integer>::Limb,
     radix_shift: <Z::Int as Integer>::Limb,
-    digits: usize,
+    digits: u32,
 ) -> DenseVector<Z> {
     let mut representative = integer.canonical();
-    let mut pieces = Vec::<Z>::with_capacity(digits);
+    let mut pieces = Vec::<Z>::with_capacity(digits as usize);
     for _ in 0..digits {
         let piece = Z::with_limb(representative & radix_mask);
         pieces.push(piece);
@@ -72,9 +72,9 @@ pub fn decompose_polynomial<Z: IntegerModRing, R: PolynomialRing<Z> + Clone>(
     polynomial: &R,
     radix_mask: <Z::Int as Integer>::Limb,
     radix_shift: <Z::Int as Integer>::Limb,
-    digits: usize,
+    digits: u32,
 ) -> DenseVector<R> {
-    let mut pieces = vec![R::ZERO; digits];
+    let mut pieces = vec![R::ZERO; digits as usize];
     decompose_impl(polynomial, radix_mask, radix_shift, &mut pieces);
     pieces.into()
 }
@@ -83,7 +83,7 @@ pub fn decompose_vector<Z: IntegerModRing, R: PolynomialRing<Z> + Clone>(
     vector: &DenseVector<R>,
     radix_mask: <Z::Int as Integer>::Limb,
     radix_shift: <Z::Int as Integer>::Limb,
-    digits: usize,
+    digits: u32,
 ) -> DenseVector<R> {
     let pieces = decompose_slice(vector, radix_mask, radix_shift, digits);
     pieces.into()
@@ -93,7 +93,7 @@ pub fn decompose_matrix<Z: IntegerModRing, R: PolynomialRing<Z> + Clone>(
     matrix: &DenseMatrix<R>,
     radix_mask: <Z::Int as Integer>::Limb,
     radix_shift: <Z::Int as Integer>::Limb,
-    digits: usize,
+    digits: u32,
 ) -> DenseMatrix<R> {
     let pieces = decompose_slice(matrix.as_ref(), radix_mask, radix_shift, digits);
     DenseMatrix::new(matrix.rows(), matrix.columns() * digits, pieces)
@@ -101,11 +101,11 @@ pub fn decompose_matrix<Z: IntegerModRing, R: PolynomialRing<Z> + Clone>(
 
 pub fn matrix<Z: IntegerModRing + Clone, R: PolynomialRing<Z> + Clone>(
     radix: &Z,
-    m: usize,
-    n: usize,
+    m: u32,
+    n: u32,
 ) -> ScalarMatrix<DenseMatrix<R>> {
     debug_assert!(n >= 2);
-    let mut powers = Vec::<R>::with_capacity(n);
+    let mut powers = Vec::<R>::with_capacity(n as usize);
     powers.push(R::ONE);
     powers.push(radix.clone().into());
     let mut power = radix.clone();
@@ -122,12 +122,12 @@ pub fn matrix<Z: IntegerModRing + Clone, R: PolynomialRing<Z> + Clone>(
 pub fn vector<Z: IntegerModRing + Clone, A: UnitalAlgebra<Z> + Clone>(
     algebra: A,
     radix: &Z,
-    digits: usize,
+    digits: u32,
 ) -> DenseVector<A>
 where
     for<'a> &'a A: AlgebraOps<Z, A>,
 {
-    let mut powers = Vec::<A>::with_capacity(digits);
+    let mut powers = Vec::<A>::with_capacity(digits as usize);
     powers.push(algebra.clone());
     let mut power = radix.clone();
     for _ in 1..digits - 1 {
